requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pillow>=9.4.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.12.0
wandb>=0.14.0
timm>=0.9.0
einops>=0.6.0
torchmetrics>=0.11.0
opencv-python>=4.7.0
albumentations>=1.3.0
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
jupyterlab>=3.6.0
ipywidgets>=8.0.0
pandas>=1.5.0
h5py>=3.8.0
beautifulsoup4>=4.11.0
requests>=2.28.0
omegaconf>=2.3.0